Initial commit
Working model of the agonist/antagonist actuator. The model has only one input: Voltage, applied to the bottom actuator. The system is perfectly symmetrical and preloaded at h0 with h0 = full height of a single actuator at rest. Alpha (sagitta of the top EBM's bulge) is solved symbolically in Wolfram Mathematica as a function of the state [h,rc] and params and its solution was put in a matlab funcion. The Lagrange equations are solved numerically. The cross section (z=0) of the system is displayed in many ways.