F(h) curves for different V levels
Computed F(h) curves at different voltage levels for the double actuator with NO reservoir. The F in the F(h) plots has a minus sign to make it positive when pointing upwards (in Lagrange it points as h downards). Checked for correct expression of eps_top in the double_EBM function and it is correct both as superposition of cone+bulge and as single eps computed from deltaR. Elastic energy has a minima in h=0, rc=ro and this can be checked imposing h_vec = [-1.5e-3, 1.5e-3], so the rest position is actually a stable equilibrium of the system when rc = ro fixed.
F(h) curves make sense, see notes.