Kapton characterization and model fit
Tests for fitting the cubic curve of the pure mechanical response of the ebm. Using both Hadi's triple pre-stretched ebms data (divide displ. by 3) and a single ebm built by myself. The young modulus of the kapton that fits the data with the model is much much lower due to many phenomenon: creases on the kapton, slippage of the kapton due to soft glue (rings), VHB stiffness not accounted for etc.